pyproject.toml
[project]
name = "hopper-gen"
version = "0.1.0"
description = "Fine-tune SDXL to paint like Edward Hopper"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diffusers>=0.30.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "peft>=0.12.0",
    "bitsandbytes>=0.43.0",
    "torch>=2.3.0",
    "modal>=0.64.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
]
